Create DarkPatterns.OpenApiCodegen.CSharp
