+++ examples/dotnetcore-server-interfaces/tooling/server-generator-references.props
@@ -4,9 +4,9 @@
 	</PropertyGroup>
 
 	<ItemGroup Condition=" '$(UseProjectReferences)' != 'true' ">
-		<PackageReference Include="DarkPatterns.OpenApiCodegen.Server.Mvc" Condition=" '$(OpenApiMvcServerFullVersion)' != '' " Version="$(OpenApiMvcServerFullVersion)" />
-		<PackageReference Include="DarkPatterns.OpenApiCodegen.Server.Mvc" Condition=" '$(OpenApiMvcServerFullVersion)' == '' and '$(VersionSuffix)' == '' " Version="$(OpenApiMvcServerVersion)" />
-		<PackageReference Include="DarkPatterns.OpenApiCodegen.Server.Mvc" Condition=" '$(OpenApiMvcServerFullVersion)' == '' and '$(VersionSuffix)' != '' " Version="$(OpenApiMvcServerVersion)-$(VersionSuffix)" />
+		<PackageReference Include="DarkPatterns.OpenApiCodegen.CSharp" Condition=" '$(OpenApiMvcServerFullVersion)' != '' " Version="$(OpenApiMvcServerFullVersion)" />
+		<PackageReference Include="DarkPatterns.OpenApiCodegen.CSharp" Condition=" '$(OpenApiMvcServerFullVersion)' == '' and '$(VersionSuffix)' == '' " Version="$(OpenApiCSharpVersion)" />
+		<PackageReference Include="DarkPatterns.OpenApiCodegen.CSharp" Condition=" '$(OpenApiMvcServerFullVersion)' == '' and '$(VersionSuffix)' != '' " Version="$(OpenApiCSharpVersion)-$(VersionSuffix)" />
 
 		<PackageReference Include="DarkPatterns.OpenApiCodegen.Json.Extensions" Condition=" '$(OpenApiJsonExtensionsFullVersion)' != '' " Version="$(OpenApiJsonExtensionsFullVersion)" />
 		<PackageReference Include="DarkPatterns.OpenApiCodegen.Json.Extensions" Condition=" '$(OpenApiJsonExtensionsFullVersion)' == '' and '$(VersionSuffix)' == '' " Version="$(OpenApiJsonExtensionsVersion)" />
@@ -16,7 +16,7 @@
 		<ProjectReference Include="$(SolutionRoot)lib\OpenApiCodegen.Json.Extensions\OpenApiCodegen.Json.Extensions.csproj" />
 
 		<ProjectReference
-			Include="$(SolutionRoot)generators\dotnetcore-server-interfaces\OpenApiCodegen.Server.Mvc.Analyzers\OpenApiCodegen.Server.Mvc.Roslyn3.11.csproj"
+			Include="$(SolutionRoot)generators\dotnetcore-server-interfaces\OpenApiCodegen.CSharp.Analyzers\OpenApiCodegen.CSharp.Roslyn3.11.csproj"
 			OutputItemType="Analyzer"
 			ReferenceOutputAssembly="false" />
 		<!-- Not an analyzer, but the project-reference analyzer needs all child dependencies, too -->
@@ -28,6 +28,6 @@
 
 	<!-- When using project references, props/targets that are in the NuGet package don't get automatically picked up -->
 	<Import Condition=" '$(UseProjectReferences)' == 'true' "
-		Project="$(SolutionRoot)generators\dotnetcore-server-interfaces\OpenApiCodegen.Server.Mvc\OpenApiCodegen.Server.Mvc.props" />
+		Project="$(SolutionRoot)generators\dotnetcore-server-interfaces\OpenApiCodegen.CSharp\OpenApiCodegen.CSharp.props" />
 
 </Project>
