+++ generators/SourceGenerators.Test/DiagnosticsShould.cs
