+++ generators/csharp/OpenApiCodegen.CSharp.Analyzers/AnalyzerReleases.Shipped.md
