+++ generators/csharp/OpenApiCodegen.CSharp.Analyzers/OpenApiCSharpGenerator.cs
@@ -1,0 +1,40 @@
+﻿using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
+using System;
+using System.Linq;
+
+#if NETSTANDARD2_0
+#nullable disable warnings
+#endif
+
+namespace DarkPatterns.OpenApiCodegen.CSharp;
+
+[Generator]
+public sealed class OpenApiCSharpGenerator() : BaseGenerator("DarkPatterns.OpenApi.CSharp.CSharpGenerator", "DarkPatterns.OpenApiCodegen.CSharp.Base")
+{
+	private const string sourceItemGroupKey = "SourceItemGroup";
+	private static readonly DiagnosticDescriptor IncludeDependentDll = new DiagnosticDescriptor(id: "DPDAPI001",
+																								title: "Include a reference to DarkPatterns.OpenApiCodegen.Json.Extensions",
+																								messageFormat: "Include a reference to DarkPatterns.OpenApiCodegen.Json.Extensions",
+																								category: "DarkPatterns.OpenApiCodegen.CSharp",
+																								DiagnosticSeverity.Warning,
+																								isEnabledByDefault: true);
+
+	protected override void ReportCompilationDiagnostics(Compilation compilation, CompilerApis apis)
+	{
+		// check that the users compilation references the expected library
+		if (!compilation.ReferencedAssemblyNames.Any(static ai => ai.Name.Equals("DarkPatterns.OpenApiCodegen.Json.Extensions", StringComparison.OrdinalIgnoreCase)))
+		{
+			apis.ReportDiagnostic(Diagnostic.Create(IncludeDependentDll, Location.None));
+		}
+	}
+
+	protected override string[] GetFileTypes(AdditionalTextWithOptions additionalText)
+	{
+		var type = additionalText.ConfigOptions.GetAdditionalFilesMetadata(sourceItemGroupKey);
+		if (type == null) return [];
+
+		// sourceGroup or sharedSourceGroup
+		return [type];
+	}
+}
