+++ generators/csharp/OpenApiCodegen.CSharp.Analyzers/OpenApiCodegen.CSharp.Roslyn3.11.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<AssemblyName>$(RootNamespacePrefix).OpenApiCodegen.Server.Mvc.Roslyn3_11</AssemblyName>
+		<AssemblyName>$(RootNamespacePrefix).OpenApiCodegen.CSharp.Roslyn3_11</AssemblyName>
 
 		<AnalyzerRoslynVersion>3.11</AnalyzerRoslynVersion>
 	</PropertyGroup>
