+++ generators/csharp/OpenApiCodegen.CSharp.Analyzers/OpenApiCodegen.CSharp.Roslyn4.0.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<AssemblyName>$(RootNamespacePrefix).OpenApiCodegen.Server.Mvc.Roslyn4_0</AssemblyName>
+		<AssemblyName>$(RootNamespacePrefix).OpenApiCodegen.CSharp.Roslyn4_0</AssemblyName>
 
 		<AnalyzerRoslynVersion>4.0</AnalyzerRoslynVersion>
 	</PropertyGroup>
