+++ generators/csharp/OpenApiCodegen.CSharp.Analyzers/OpenApiCodegen.CSharp.Roslyn4.1.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<AssemblyName>$(RootNamespacePrefix).OpenApiCodegen.Server.Mvc.Roslyn4_1</AssemblyName>
+		<AssemblyName>$(RootNamespacePrefix).OpenApiCodegen.CSharp.Roslyn4_1</AssemblyName>
 
 		<AnalyzerRoslynVersion>4.1</AnalyzerRoslynVersion>
 	</PropertyGroup>
