+++ generators/csharp/OpenApiCodegen.CSharp.Analyzers/OpenApiCodegen.CSharp.Roslyn4.1.props
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Project>
+	<Import Project="../OpenApiCodegen.CSharp/OpenApiCodegen.CSharp.props" />
+</Project>
