+++ generators/csharp/OpenApiCodegen.CSharp.Analyzers/Properties/launchSettings.json
