+++ generators/csharp/OpenApiCodegen.CSharp.Base/CSharpSchemaServerExtensionsOptions.cs
