+++ generators/csharp/OpenApiCodegen.CSharp.Base/CSharpServerSchemaOptions.cs
