+++ generators/csharp/OpenApiCodegen.CSharp.Base/ControllerHandlebarsTemplateProcess.cs
