+++ generators/csharp/OpenApiCodegen.CSharp.Base/Templates/AddServices.cs
