+++ generators/csharp/OpenApiCodegen.CSharp.Base/Templates/addServices.handlebars
