+++ generators/csharp/OpenApiCodegen.CSharp.Base/Templates/controller.handlebars
