+++ generators/csharp/OpenApiCodegen.CSharp.Base/csharp.config.yaml
