+++ generators/csharp/OpenApiCodegen.CSharp/OpenApiCodegen.CSharp.csproj
@@ -7,11 +7,10 @@
 
 		<IsPackable>true</IsPackable>
 		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-		<IsRoslynComponent>true</IsRoslynComponent>
 
-		<title>Controller source generators from OpenAPI specification</title>
-		<PackageSummary>Adds source generators to generate a C# client (via extension methods) from an OpenAPI specification file.</PackageSummary>
-		<description>Adds source generators to generate a C# client (via extension methods)  from an OpenAPI specification file.</description>
+		<title>C# source generators from OpenAPI specification</title>
+		<PackageSummary>Adds source generators to generate C# classes from an OpenAPI specification file.</PackageSummary>
+		<description>Adds source generators to generate C# classes from an OpenAPI specification file, including abstract controllers and csharp clients.</description>
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
 		<IncludeBuildOutput>false</IncludeBuildOutput>
 		<CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
@@ -23,11 +22,10 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<AnalyzerReference Include="..\OpenApiCodegen.Client.Analyzers\OpenApiCodegen.Client.Roslyn3.11.csproj" />
-		<AnalyzerReference Include="..\OpenApiCodegen.Client.Analyzers\OpenApiCodegen.Client.Roslyn4.0.csproj"
+		<AnalyzerReference Include="..\OpenApiCodegen.CSharp.Analyzers\OpenApiCodegen.CSharp.Roslyn3.11.csproj"
 						   ReferenceAnalyzer="false" />
-		<ProjectReference Include="$(SolutionRoot)generators\dotnetstandard-client\OpenApiCodegen.Client.Base\OpenApiCodegen.Client.Base.csproj" SpecificVersion="true"
-						  ReferenceOutputAssembly="false" AnalyzerDynamicContext="true" />
+		<AnalyzerReference Include="..\OpenApiCodegen.CSharp.Analyzers\OpenApiCodegen.CSharp.Roslyn4.0.csproj"
+						   ReferenceAnalyzer="false" />
 	</ItemGroup>
 
 	<Target Name="AddDlls" BeforeTargets="_GetPackageFiles">
