+++ generators/csharp/OpenApiCodegen.CSharp/OpenApiCodegen.CSharp.props
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
+<?xml version="1.0" encoding="utf-8" ?>
 <Project>
 	<ItemGroup>
 		<CompilerVisibleProperty Include="RootNamespace" />
