--- generators/dotnetcore-server-interfaces/OpenApiCodegen.Server.Mvc.Analyzers/OpenApiCodegen.Server.Mvc.Roslyn4.1.props
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Project>
-	<Import Project="../OpenApiCodegen.Server.Mvc/ControllerGenerator.props" />
-</Project>
