--- generators/dotnetcore-server-interfaces/OpenApiCodegen.Server.Mvc/OpenApiCodegen.Server.Mvc.props
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Project>
-	<Import Project="ControllerGenerator.props" />
-</Project>
