+++ lib/OpenApiCodegen.Server.Mvc.Test/OpenApiCodegen.Server.Mvc.Test.csproj
@@ -26,7 +26,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="$(SolutionRoot)generators\dotnetcore-server-interfaces\OpenApiCodegen.Server.Mvc.Base\OpenApiCodegen.Server.Mvc.Base.csproj" />
+		<ProjectReference Include="$(SolutionRoot)generators\csharp\OpenApiCodegen.CSharp.Base\OpenApiCodegen.CSharp.Base.csproj" />
 		<ProjectReference Include="..\OpenApiCodegen.Handlebars\OpenApiCodegen.Handlebars.csproj" />
 		<ProjectReference Include="..\OpenApiCodegen.Json.Extensions\OpenApiCodegen.Json.Extensions.csproj" />
 		<ProjectReference Include="..\OpenApiCodegen.TestUtils\OpenApiCodegen.TestUtils.csproj" />
