+++ lib/TestApp/OpenApiCodegen.Server.Mvc.TestApp.csproj
@@ -37,13 +37,13 @@
 
   <ItemGroup>
     <ProjectReference
-      Include="$(SolutionRoot)generators\dotnetcore-server-interfaces\OpenApiCodegen.Server.Mvc.Analyzers\OpenApiCodegen.Server.Mvc.Roslyn4.1.csproj"
+      Include="$(SolutionRoot)generators\csharp\OpenApiCodegen.CSharp.Analyzers\OpenApiCodegen.CSharp.Roslyn4.0.csproj"
       OutputItemType="Analyzer"
       ReferenceOutputAssembly="false" />
     <ProjectReference Include="..\OpenApiCodegen.Json.Extensions\OpenApiCodegen.Json.Extensions.csproj" />
   </ItemGroup>
 
   <!-- The project reference won't include this due to the multi-version targeting trick -->
-  <Import Project="$(SolutionRoot)generators\dotnetcore-server-interfaces\OpenApiCodegen.Server.Mvc\OpenApiCodegen.Server.Mvc.props" />
+  <Import Project="$(SolutionRoot)generators\csharp\OpenApiCodegen.CSharp\OpenApiCodegen.CSharp.props" />
 
 </Project>
